backend/requirements.txt
asgiref==3.11.1
brotli==1.2.0
dj-database-url==2.1.0
Django==4.2.28
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==21.2.0
packaging==26.0
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.0.1
qrcode==8.2
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.6.0
